design fixes for trade and component restructure
