+++ src/components/cardList/styles.ts
@@ -1,26 +1,6 @@
 import styled, { css } from 'styled-components'
 import { mobile, sdesktop, tablet } from 'themes/screens'
 
-export const PageContainer = styled.div`
-  margin: 0px auto 20px auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  padding: 0px 50px 50px 50px;
-  width: 100%;
-  max-width: 1164px;
-  ${sdesktop(css`
-    padding: 0px 0px 50px 0px;
-  `)}
-  ${tablet(css`
-    padding: 0px 50px;
-  `)}
-  ${mobile(css`
-    padding: 0px 25px;
-    margin: auto;
-  `)}
-`
-
 export const CardList = styled.div`
   display: grid;
   gap: 15px;
@@ -108,6 +88,10 @@ export const Title = styled.a`
   &:hover {
     text-decoration: underline;
   }
+
+  ${tablet`
+   font-size: 14px;
+  `}
 
   ${(props) => css`
     color: ${props.theme.name === 'light'
